+++ src/api/auth/oauth.api.ts
@@ -5,9 +5,22 @@ export interface SocialoginCompleteReq {
   provider: string;
   code: string;
 }
-export type EmptyObject = Record<string, never>;
+export type SocialProfile = {
+  provider: "KAKAO" | string;
+  providerId: string;
+  name: string;
+  nickname: string;
+  profileImage: string;
+};
 
-export type SocialoginCompleteData = EmptyObject; // 나중에 타입 나오면 정의하기
+export type SocialoginCompleteData = {
+  isNewMember: boolean;
+  signupToken?: string;
+  socialProfile?: SocialProfile;
+  // 만약 기존 회원이면 백이 access/refresh를 줄 수도 있으니 확장 가능
+  accessToken?: string;
+  refreshToken?: string;
+};
 
 //소셜 로그인 (웹 브라우저용)
 export async function oauth(
